Modularises event modal
